# Major change: Timeline logic takes conflicts into account in references