--- a/.eslintignore
+++ b/.eslintignore
@@ -1,1 +1,2 @@
 examples/*.js
+tests/*.js