--- a/.gitignore
+++ b/.gitignore
@@ -12,3 +12,4 @@ wallaby.conf.js
 
 .DS_Store
 .vscode/settings.json
+tests/