--- a/src/api/enums.ts
+++ b/src/api/enums.ts
@@ -1,5 +1,0 @@
-export enum EventType {
-	START = 0,
-	END = 1,
-	KEYFRAME = 2,
-}