--- a/src/api/expression.ts
+++ b/src/api/expression.ts
@@ -1,0 +1,12 @@
+export type Expression = number | string | ExpressionObj | null
+
+export type ExpressionOperator = '+' | '-' | '*' | '/' | '&' | '|' | '!' | '%'
+export interface ExpressionObj {
+	l: Expression
+	o: ExpressionOperator
+	r: Expression
+}
+export interface InnerExpression {
+	inner: Array<any> // string or Array<string>
+	rest: Array<string>
+}