--- a/src/api/expression.ts
+++ b/src/api/expression.ts
@@ -1,3 +1,8 @@
+/** An Expression can be:
+ * * An absolute value (number)
+ * * An expression describing a relationship (string), like "#A.start + 10"
+ * * An expression object, like {l: "#A.start", o: '+', r: '10'}
+ */
 export type Expression = number | string | ExpressionObj | null
 
 export type ExpressionOperator = '+' | '-' | '*' | '/' | '&' | '|' | '!' | '%'