--- a/src/api/index.ts
+++ b/src/api/index.ts
@@ -1,0 +1,6 @@
+export * from './expression'
+export * from './resolvedTimeline'
+export * from './resolver'
+export * from './state'
+export * from './timeline'
+export * from './types'