--- a/src/api/resolvedTimeline.ts
+++ b/src/api/resolvedTimeline.ts
@@ -26,10 +26,8 @@ export interface ResolvedTimeline<TContent extends Content = Content> {
 	nextEvents: Array<NextEvent>
 
 	statistics: {
-		/** Number of objects that were unable to resolve */
-		unresolvedCount: number
-		/** Number of objects that were resolved */
-		resolvedCount: number
+		/** Number of timeline objects (including keyframes) in the timeline */
+		totalCount: number
 
 		/** Number of resolved instances */
 		resolvedInstanceCount: number
@@ -40,7 +38,15 @@ export interface ResolvedTimeline<TContent extends Content = Content> {
 		/** Number of resolved keyframes */
 		resolvedKeyframeCount: number
 
-		/** How many objects that was actually resolved (is affected when using cache) */
+		/** How many objects (including keyframes) where actually resolved (is affected when using cache) */
+		resolvingObjectCount: number
+
+		/**
+		 * How many times the objects where resolved.
+		 * If there are deep/recursive references in the timline,
+		 * the resolver might resolve an object multiple times in order to fully resolve the timeline.
+		 * (is affected when using cache)
+		 */
 		resolvingCount: number
 	}
 }