--- a/src/api/state.ts
+++ b/src/api/state.ts
@@ -1,0 +1,11 @@
+import { NextEvent, ResolvedTimelineObjectInstance } from './resolvedTimeline'
+import { Time } from './types'
+
+export interface TimelineState {
+	time: Time
+	layers: StateInTime
+	nextEvents: NextEvent[]
+}
+export interface StateInTime {
+	[layer: string]: ResolvedTimelineObjectInstance
+}