--- a/src/resolver/__tests__/ResolvedTimelineHandler.spec.ts
+++ b/src/resolver/__tests__/ResolvedTimelineHandler.spec.ts
@@ -1,0 +1,2 @@
+import { ResolvedTimelineHandler } from '../ResolvedTimelineHandler'
+import { baseInstances } from '../lib/instance'