… .github/workflows/node.yml
… .github/workflows/publish-prerelease.yml
… .github/workflows/sonar.yml
… MIGRATION.md
… README.md
… examples/simple.js
… examples/simple.ts
… package.json
… src/__tests__/basic.spec.ts
… src/__tests__/expression.spec.ts
… src/__tests__/groups.spec.ts
… src/__tests__/index.spec.ts
… src/__tests__/invalidate.spec.ts
… src/__tests__/keyframes.spec.ts
… src/__tests__/legacy.spec.ts
… src/__tests__/lib.spec.ts
… src/__tests__/performance.spec.ts
… src/__tests__/validate.spec.ts
… src/api/api.ts
… src/api/timeline.ts
… src/api/types.ts
… src/index.ts
… src/lib.ts
… src/resolver/CacheHandler.ts
… src/resolver/ExpressionHandler.ts
… src/resolver/InstanceHandler.ts
… src/resolver/LayerStateHandler.ts
… src/resolver/ReferenceHandler.ts
… src/resolver/ResolvedTimelineHandler.ts
… src/resolver/ResolverHandler.ts