This is a big rewrite of the timeline library.

### Summary of changes

* Large refactoring, moving internal functions into classes
* Change resolving from 2 steps (`resolveTimeline()`, `resolveAllStates()`) to just 1 (`resolveTimeline()`)
* **Main change:** Change how the timeline logic works, se _Timeline logic change_ below.

### Test it

```
npm install superfly-timeline@9.0.0-nightly-wip-big-rewrite-20230814-192948-870dbc7.0
```

* JSFiddle, old version 8: https://jsfiddle.net/nytamin/r1L48hu3/
* JSFiddle, new version 9: https://jsfiddle.net/nytamin/rxkf2p8L/
  
### TODO

- [x] Make all unit tests pass.
      Status: 2023-06-16: 199 / 273 (72%)
      Status: 2023-06-19: 226 / 290 (77%)
      Status: 2023-06-21: 234 / 295 (79%)
      Status: 2023-06-28: 295 / 301 (98%)
      Status: 2023-06-30: 301 / 301 (100%)
- [x] Re-implement existing functionality
- [x] clean up some code (make sonarCloud reasonably happy)
- [x] Add documentation for methods and classes
- [x] Add documentation inside of some methods
- [x] Update examples
- [x] Document which pre-existing unit-tests that fails, and how they where rewritten to account for new timeline logic
- [x] Squash & tidy up git commit log
- [x] Request code reviews
- [x] Update Readme
- [x] Publish nightly to npm for testing
- [x] Look into previously unsovable PR:s:
  - [x] #52 
  - [x] #92
- [x] Code Review
- [x] (After summer) Merge PR

# Migration notes

### API change
- The `Resolver` namespace has been removed. Its methods are now exported as separate functions instead.
- `Resolver.resolveTimeline()` and `Resolver.resolveAllStates()` have been combined into one: `resolveTimeline()`
- `Resolver.getState` has been renamed to `getResolvedState`
- `validateIdString` has been renamed to `validateReferenceString`.
- `resolvedTimeline.statistics` properties have changed.


```typescript
// Before

// Resolve the timeline
const options: ResolveOptions = {
	time: 0,
}
const resolvedTimeline = Resolver.resolveTimeline(timeline, options)
const resolvedStates = Resolver.resolveAllStates(resolvedTimeline)
// Calculate the state at a certain time:
const state = Resolver.getState(resolvedStates, 15)


// After

// Resolve the timeline
const options: ResolveOptions = {
	time: 0,
}
const resolvedTimeline = resolveTimeline(timeline, options)
// Calculate the state at a certain time:
const state = getResolvedState(resolvedTimeline, 15)
```

### Timeline logic change

Before, references where evaluated on the original (non conflicted timeline objects).
After, the references are updated when a conflict affects the dependees.

```typescript
const timeline = {
    {id: 'A', layer: '1', enable: {start: 10, end: 100}}
    {id: 'B', layer: '1', enable: {start: 50, end: null}}

    {id: 'X', layer: '1', enable: {while: '#A'}}
}

// Before:
// A playing at [{start: 10, end: 50 }] (interrupted by B)
// B playing at [{start: 50, end: null }]
// X playing at [{start: 10, end: 100 }] (still references the original times of A)

// After:
// A playing at [{start: 10, end: 50 }] (interrupted by B)
// B playing at [{start: 50, end: null }]
// X playing at [{start: 10, end: 50 }] (references the updated times of A)
```

### About performance

The new timeline logic is heavier to resolve for timelines where objects reference objects that are affected by collisions.
For timelines where there are no such relationships, resolving is actually slightly faster thanks to some added optimizations.

Performance results:
* Static tests (`yarn unit performance`): Without cache 8.4 ms (before: 5.5ms), with cache: 5.80ms ( before 4ms)
* Typical real-world [NRK Sofie](https://github.com/nrkno/sofie-core) timelines: 
  * A (141 objects): Without cache 14.85ms (before: 15ms )
  * B (112 objects): Without cache 4.20ms (before: 4.50ms )